+++ tabletest-junit/src/test/java/io/github/nchaugen/tabletest/junit/features/JavaNullValueTest.java
@@ -1,34 +1,29 @@
 package io.github.nchaugen.tabletest.junit.features;
 
-import io.github.nchaugen.tabletest.junit.Scenario;
 import io.github.nchaugen.tabletest.junit.TableTest;
 
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
+import static io.github.nchaugen.tabletest.junit.TableTestExceptionAssertions.assertThrowsWhenFallbackFails;
 import static io.github.nchaugen.tabletest.junit.TableTestExceptionAssertions.assertThrowsWhenNullSpecifiedForPrimitiveType;
-import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNull;
 
 class JavaNullValueTest {
 
     @TableTest("""
-        Scenario            | String | Integer | List | Map | Set
-        Blank               |        |         |      |     |
-        Empty single quoted | ''     | ''      | ''   | ''  | ''
-        Empty double quoted | ""     | ""      | ""   | ""  | ""
+        String | Integer | List | Map | Set
+               |         |      |     |
         """)
-    void blank_is_null_and_empty_is_null_for_non_string(
-        @Scenario String scenario,
+    void blank_converts_to_null(
         String string,
         Integer integer,
         List<?> list,
         Map<String, ?> map,
         Set<?> set
     ) {
-        if ("Blank".equals(scenario)) assertNull(string);
-        else assertEquals("", string);
+        assertNull(string);
         assertNull(integer);
         assertNull(list);
         assertNull(map);
@@ -38,10 +33,24 @@ class JavaNullValueTest {
     @TableTest("""
         table value | parameter type
                     | boolean
-        ''          | int
+                    | short
+                    | char
         """)
-    void fails_when_null_value_specified_for_primitive_parameter_type(String value, Class<?> type) {
+    void blank_fails_for_primitive_parameter_type(String value, Class<?> type) {
         assertThrowsWhenNullSpecifiedForPrimitiveType(value, type);
+    }
+
+    @TableTest("""
+        Scenario            | table value | Type
+        Empty single quoted | ''          | java.util.List
+        Empty double quoted | ""          | java.lang.Integer
+        Empty primitive     | ""          | boolean
+        """)
+    void empty_string_for_non_string_types_requires_factory_method(
+        String value,
+        Class<?> type
+    ) {
+        assertThrowsWhenFallbackFails(value, type);
     }
 
 }