+++ tabletest-junit/src/test/kotlin/io/github/nchaugen/tabletest/junit/converting/KotlinConversionFailureTest.kt
@@ -17,11 +17,8 @@ class KotlinConversionFailureTest {
         52          | io.github.nchaugen.tabletest.junit.javadomain.Ages
             """
     )
-    fun fails_conversion_for_single_values_outside_type_range(
-        value: String,
-        type: Class<*>
-    ) {
-        assertThrowsWhenFallbackFails(value, type.getTypeName())
+    fun fails_conversion_for_single_values_outside_type_range(value: String, type: Class<*>) {
+        assertThrowsWhenFallbackFails(value, type)
     }
 
     @TableTest(